Clean up some code and a small attempt at FAT12 support
